Make Calendar widget treat multi-day events the same as allDay.
Events that start mid-day and span multiple days are not assigned the allDay flag, and were being treated as normal events.  This caused us to show the same event on the calendar widget for several days, until half-way through the multi-day event.

This change treats any events with duration longer than a single day as allDay events, which triggers a widget change at the event start instead of half-way through the event.

http://b/issue?id=1772674