Add calendar organizer to event on create.
When an event is created, the CalendarProvider should add the organizer field.
This data is client-only, and is not sent to the server.

Also change behavior of SELF_ATTENDEE_STATUS to create Attendees entry for
the owner (email in the calendar URL) rather than for the _sync_account
(the user accessing the calendar).  It appears the old behavior was
wrong for shared calendars.