b/1522903 First step in converting allday events to local timezone
This is the first change for converting allDay events to use the local
timezone instead of being in UTC timezones. This will make dealing with
alarms on allDay events easier to manage and allow us to remove a large
number of checks throughout the code to deal with allDay events
differently.