b/2412594 Changed widget to use an implicit intent.
Added an implicent intent filter to the Calendar launcher so it will no longer be tied
to the package name. The widget is now using that intent to launch the Calendar so it
doesn't have to use a hardcoded package name. Also added uri matching for a time to be
specified as part of the data field in the intent.