Make instance expansion specific to a calendar.
In the old code, canceling of recurrence instances due to exceptions was
based on the sync_id and time.  If two calendars happened to have events
with the same sync_id and time, the wrong cancellations could happen.

In this change, sync_id:calendar_id is used to track instances and cancellations
to ensure uniqueness.

bug 2425227

Change-Id: I0b79340f3f45a124f3e3850754c5e5c9cb2bd5a0