b/3008125 Make CP update the default prevTZ
The phone will default to GMT as the time zone on a fresh wipe
which is not a time zone that Calendar officially supports. This
change makes it so when the phone figures out where it is the
provider will update the default previous time zone to the phone's
new location instead of using GMT. This was manually merged from
gb.

Change-Id: I8d641b92ed488f243119fdb1fdd194b16b7e151f