Prepare Calendar Provider for GData 2.0 & Partial Update
This CL prepares support for GData 2.0 by adding the _SYNC_7  column to the Events table.
This column will contain the edit URL which used to be in SYNC_VERSION but in 2.0, will hold the eTag instead.

It also prepares for Partial Update by adding a column called lastSynced to Events & a column called canPartiallyUpdate to Calendars.
If a sync adapter sets canPartiallyUpdate to 1, the provider will dup events before modifying them locally and mark them as dups using the lastSynced column.
When a sync adapter then updates an event and sets DIRTY to 0, the provider will remove the duplicated event.

Since no sync adapters yet set canPartiallyUpdate or use SYNC_7, this change should effectivly be a NOP.

Change-Id: I651056b3c0306391e9e559ceb185e77187118846