Feature + exception fixes in CalendarProvider
New feature: if selfAttendeeStatus is updated when creating an
exception for a recurring event, the provider will automatically update
the "status" columns in the appropriate Attendees rows.  (Normally this
goes the other direction, where updates to the Attendees table are
mirrored into Events.)

Added some test code to exercise it (though we're still not verifying
changes to Attendees).

Fix: the exception creation code was mis-handling the start date.  This
didn't show up in testing because the tests were creating exceptions
with modified start dates.

Fix: the tests now fiddle with the duration rather than the start date,
so that problems like the above are visible.

Fix: don't allow DTEND to be specified in exception creation.  This
value is automatically generated from originalInstanceTime + duration.
Also added two recent SYNC_DATA fields to the "don't copy" list.

Fix: set the dirty flag on exception events so the server will see them
(unless the creator is the sync adapter).  Note the CTS tests do most
of their work claiming to be the sync adapter, so this doesn't really
show up there.

Bug 2664212

Change-Id: Iab43171926ee51a5ba0f9193830aa97b9b4fc147