Avoid re-use of orphaned data
Data is being added to Attendees and Reminders that is associated
with delete Events.  When the Event ID is re-used, the orphaned
data attaches to the new event.  To avoid this, change the Event
ID column to "autoincrement", ensuring that IDs are never re-used.

At database upgrade time, delete any orphaned rows.

Bug 5063315

Change-Id: I02051bf271ea9b9fd05c27e50f546177f42f43d0