DO NOT MERGE - Avoid re-use of orphaned data
Data is being added to Attendees and Reminders that is associated
with delete Events.  When the Event ID is re-used, the orphaned
data attaches to the new event.  To avoid this, change the Event
ID column to "autoincrement", ensuring that IDs are never re-used.

At database upgrade time, delete any orphaned rows.

Bug 5063315

Change-Id: Ic2519b335d6d6b2c82579c5484ad3b8528ffd682