Scrub some, validate others
The previous behavior of the Calendar provider was to scrub all
new events.  This was changed to scrub and validate all new events
and all updates.  This now changes it again: we now scrub new events
and updates that originate from a sync adapter, and validate new
events and updates that come from an application.

The idea is to retain compatibility with the sync adapters while
clamping down on apps.

Bug 2519655

Change-Id: I792ed1db126947ae99a6e97f81e9f2380dfa1a7f