Don't re-use scrubbed mod values
The "scrub" code was changed so that it added explicit null
assignments to the values-to-be-updated set.  This is fine for
single updates, but for multi-event updates (which happen during
the initial sync) we were nulling out stuff in later events.  With
this change we make a copy of the set before modifying it.

Bug 2519655

Change-Id: Ib62a267bf6a45b57c85685e439ae4100487dd9df