Update provider for EMMA from CTS
The provider now accepts a "secret" request to dump EMMA coverage
data.

The commands in tests/EMMA_README were updated for the new world
order, in which the tests live in CTS rather than being one with
the provider itself.  This also converts the commands from direct
execution to a shell function, and adds a second function that can
be used to de-EMMA the provider.

Bug 2361048

Change-Id: I38b494cc65a15523207f95d2e622756cc393b665