Add URIs for updating Attendees and Reminders
Previously you could only update a single entry in Attendees or
Reminders, using the "/#" form of the URI.  Now you can update
multiple entries by specifying a selection.  (This is not something
apps will commonly want to do, but if nothing else it rounds out
the APIs.)

Also, merged common code for updates to Attendees, Reminders, and
ExtendedProperties.

Also, fixed a couple of minor bugs with updating an Attendee and
the associated Event.selfAttendeeStatus (e.g. event_id had to be
specified in the update values or the provider would NPE).

Bug 5373449

Change-Id: I72df47ece69bf083739970ffa48a213a2cf76704