Remove SyncCalendarTest
We have CRUD test in CalendarProvider2Test.java.
This test case is to test whether the sync adapter work after the CRUD
operation. Tests should not be in here and should rather in
CalendarSyncAdapter test project. And CalendarSyncAdapter has its own
testing already and they are far more robust than what we have here.

Also, it requires the testing device to have a google account
which is actually not make sense in CalendarProvider test project.

This is the final piece that to make all tests in CalendarProvider test
project all pass. I will add the test project to continuous integartion.

Bug: 21689286
Change-Id: I61668f43b93cf8f98c02390eb16e709736affe82