Remove SCHEDULE_EXACT_ALARM permission
The permission is user visible and can cause calendar apps to
malfunction if revoked, even if the calendar apps themselves have this
permission. This is a system component and not user visible, so we
should do our best to not confuse the user by showing this to them as a
separate "app".

Upgrading the allow-list should let the provider use exact alarms while
the device is in doze.

Test: Manually navigate to:
Settings -> Apps -> Special app access -> Alarms & Reminders -> 3-dot
menu -> Show system
No entry should be shown as "Calendar Storage".
logcat should show no crashes for "com.android.providers.calendar"

Bug: 171306433
Change-Id: I5c7ce5a6e0accaf05520b326b0eb956013393d10