Update the handling of BOOT_COMPLETED broadcast within CalendarProvider.
CalendarProvider's post-initialization work can potentially lead to
contention with its BOOT_COMPLETED work due to both of them accessing
the DB and performing similar operations. Although this is rare, it
could lead to an ANR since the broadcast isn't handled in time.

Defer the BOOT_COMPLETED work which can cause contention to after the
post-initialization work is completed, ensuring both tasks are always
executed.

If the provider is out of lifecycle by the time it gets to handling of
the deferred broadcast work, the deferred work will be executed
eventually when any operation is performed on the provider.

Bug: 388910023
Test: atest CalendarProviderTests
Flag: com.android.providers.calendar.defer_post_initialize_work
Change-Id: I4674140cee3d157631e6a908fbfab911b87ca071